requirements.txt
selenium==4.10.0
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
websockets==11.0.3
jinja2==3.1.2
aiofiles==23.1.0
pydantic==2.4.2
python-dotenv==1.0.0
webdriver-manager==4.0.0
pillow==10.0.0